+++ mmd_tools/operators/display_item.py
@@ -226,7 +226,7 @@ class SelectCurrentDisplayItem(Operator):
                 mmd_root.active_morph_type = item.morph_type
                 mmd_root.active_morph = index
         else:
-            utils.selectSingleBone(context, mmd_model.Model(root).armature(), item.name)
+            utils.selectSingleBone(context, mmd_model.FnModel.find_armature(root), item.name)
         return {'FINISHED'}
 
 @register_wrap
