+++ mmd_tools/panels/prop_object.py
@@ -73,7 +73,7 @@ class MMDRigidPanel(_PanelBase, Panel):
             row.prop(obj.mmd_rigid, 'bone', text='', icon='BONE_DATA')
         else:
             row = c.row(align=True)
-            armature = mmd_model.Model(root).armature()
+            armature = mmd_model.FnModel.find_armature(root)
             row.prop_search(obj.mmd_rigid, 'bone', text='', search_data=armature.pose, search_property='bones', icon='BONE_DATA')
 
         c = layout.column(align=True)
