+++ mmd_tools/properties/__init__.py
@@ -20,7 +20,6 @@ else:
         material,
         bone,
         rigid_body,
-        armature,
         )
 
 __properties = {
@@ -53,7 +52,7 @@ __properties = {
         'mmd_rigid': bpy.props.PointerProperty(type=rigid_body.MMDRigidBody),
         'mmd_joint': bpy.props.PointerProperty(type=rigid_body.MMDJoint),
         'is_mmd_glsl_light': bpy.props.BoolProperty(name='is_mmd_glsl_light', default=False),
-        'mmd_data_query': bpy.props.PointerProperty(type=armature.MMDDataQuery),
+        'mmd_data_query': bpy.props.PointerProperty(type=root.MMDDataQuery),
         },
     bpy.types.Material: {
         'mmd_material': bpy.props.PointerProperty(type=material.MMDMaterial),
